+++ resources/android/com/theunwindfront/audio/AudioService.kt
@@ -13,6 +13,7 @@ import android.os.IBinder
 import androidx.core.app.NotificationCompat
 import androidx.core.app.ServiceCompat
 import androidx.media.app.NotificationCompat.MediaStyle
+import android.support.v4.media.session.MediaControllerCompat
 
 class AudioService : Service() {